watcher/async: drain mach port on notify
We must drain the mach port when we are notified that there is data
available on the port. We specify MACH_RCV_MSG for kqueue but this only
drains the first message and leaves others queued. That is reasonable
behavior! But, it is not the behavior we want.

kqueue doesn't appear to have a flag to do this so we do it manually.