iocp: fix a race condition in the process watcher
Previously, there was a race condition in the IOCP process watcher where
if the process exited before wait() was called, it would hang indefinitely.

I addressed this by adding a new result type for when a job is associated
with the completion port, and calling the callback with this result so that
the user code has a chance to check any invariants.

Changes:
- Call the completion callback when the job is associated with the IOCP
- Add a test that tests the case of the process exiting before the wait() call
- Rework job object handling to not require @unionInit. The user code
  can now interpret the message data however they want