backend/kqueue: fix empty udata crash
On certain occasions, internal kqueue events are dispatched with an empty udata field.
This causes a crash when trying to access the udata field in the event handler.

This patch adds a check to ensure that the udata field is not empty before accessing it
just like in the other events consumer.