stream: add poll function for read polling
This adds a stream `poll` function that works with files, sockets, and
other fd-based types to allow for polling. This commit only implements
read polling, but write polling will be added in a future commit.

Notably, this is useful because polling never requires a threadpool, so
if the caller wanted to handle when they read on their own, they can use
this instead of `read()` (which -- on certain backends -- does require a
threadpool).

io_uring and epoll have native `poll` event types so this is easy. For
kqueue, I had to make a special case for empty read buffers to avoid
reading and just triggering the callback as readable.