io_uring: add msg_ring functionality
io_uring has a function called msg_ring which adds a CQE to a target
ring, which may be operating in a different thread. This can be very
valuable to wakeup a sleeping thread while also giving it data to
operate with.

Zig std doesn't have the prep function from liburing yet, so we emulate
how that function works. Specifically, we are emulating
io_uring_prep_msg_ring_cqe_flags. We are using the *cqe_flags variant so
that we can set a flag on the receiver's CQE. Because of how completion
callbacks work, the same callback will be called twice: once with the
sender's loop and once with the receiver's loop. By setting a flag, we
can distinguish between these. The method allows passing a u32 value
through as extra data, in addition to the pointer. We also pass this
through, but when we detect we are the sender of the message we return a
null. Only the receiver of the message will get the value contained
here.