Swap read on timerfd/eventfd to poll to avoid Kernel 6.15.4 bug
This commit changes our operations on timerfd and eventfd from using
read to using poll. This is a workaround for a bug in Kernel 6.15.4 that
causes read operations on these file descriptors to hang indefinitely.

A poll is equivalent to what we're trying to achieve here and is
equivalent to epoll because epoll is already polling and `read` was just
a wrapper that subsequently called `read` anyways.

Prior to this commit, tests would hang on 6.15.4, now they pass.