remove usingnamespace from main.zig
This should be 100% API compatible, and I wrote a comptime test to help
ensure that. Still, I suppose something could slip through.

The major churn in this PR is to avoid a circular dependency when
verifying that the API didn't change. We remove the dependency on
main.zig from the backends, which is cleaner in general, but it makes
for a big diff.